move to utils/lang
